Fix name resolution errors when using SOCKS proxy
Calling socket.getaddrinfo() in TCPClient.create_connection with the
target domain might fail, or might resolve the wrong address if we
want to resolve names over the SOCKS proxy (dns_over_socks=true).
This commit changes this by using the SOCKS proxy as the "host"
parameter in socket.getaddrinfo() if SOCKS is used. After this, we
obviously can't use "sa" from the return value of getaddrinfo(), but
have to use the original address.